[refactor] 배송지 컴포넌트 분리 및 메인 이동 컴포넌트 추가
