+++ dutchiepay/src/app/_components/_mypage/_delivery/DeliveryAddress.jsx
@@ -6,11 +6,12 @@ import '@/styles/globals.css';
 import { useDispatch, useSelector } from 'react-redux';
 import { useEffect, useState } from 'react';
 
+import AddressAddition from '@/app/_components/_mypage/_delivery/AddressAddition';
 import CryptoJS from 'crypto-js';
-import DeliveryAddressItem from './DeliveryAddressItem';
+import DeliveryAddressItem from '@/app/_components/_mypage/_delivery/DeliveryAddressItem';
 import Image from 'next/image';
 import axios from 'axios';
-import delivery from '../../../../public/image/delivery.svg';
+import delivery from '/public/image/delivery.svg';
 import { setAddresses } from '@/redux/slice/addressSlice';
 
 export default function DeliveryAddress() {
@@ -57,24 +58,6 @@ export default function DeliveryAddress() {
     }
   }, [isChanged]);
 
-  useEffect(() => {
-    const handleMessage = (event) => {
-      if (event.origin !== window.location.origin) return;
-      if (
-        event.data &&
-        (event.data.type === 'ADD_ADDRESS' ||
-          event.data.type === 'UPDATE_ADDRESS')
-      ) {
-        setIsChanged(true);
-      }
-    };
-
-    window.addEventListener('message', handleMessage);
-    return () => {
-      window.removeEventListener('message', handleMessage);
-    };
-  }, []);
-
   return (
     <article>
       <h2 className="font-semibold text-2xl">배송지 관리</h2>
@@ -107,22 +90,10 @@ export default function DeliveryAddress() {
           ))}
         </div>
       )}
-      <button
-        className="w-[150px] mt-[24px] px-[24px] py-[8px] bg-blue--500 text-white text-sm flex justify-between rounded block mx-auto"
-        onClick={() => {
-          if (!deliveryAddress || deliveryAddress.length < 5) {
-            window.open(
-              '/delivery-address',
-              '주소지 추가',
-              'width=620, height=670, location=1'
-            );
-          } else {
-            alert('최대 5개까지의 주소지를 저장할 수 있습니다.');
-          }
-        }}
-      >
-        주소지 추가<p>+</p>
-      </button>
+      <AddressAddition
+        deliveryAddress={deliveryAddress}
+        setIsChanged={setIsChanged}
+      />
     </article>
   );
 }
