+++ dutchiepay/src/app/_components/_user/GoToMain.jsx
@@ -1,0 +1,16 @@
+import '@/styles/globals.css';
+import '@/styles/user.css';
+
+import Link from 'next/link';
+
+export default function GoToMain() {
+  return (
+    <Link
+      href="/"
+      className="mt-[24px] text-gray--500 text-sm text-center underline"
+      role="button"
+    >
+      메인으로 돌아가기
+    </Link>
+  );
+}
